requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
